WIP: zstd / float16 / int8 encodings, SIMD, HNSW, BM25, presets
Section payload encodings 1/2/3 (zstd / float16 / int8) wired through
the writer, reader, and runtime. content_hash now hashes decoded bytes
so a zstd-compressed corpus shares the same content_hash (and therefore
citation URIs) as its raw equivalent. Section checksum still hashes
physical bytes; file_hash still covers the whole file.

Runtime adds a SIMD dispatcher (AVX2 / NEON / scalar) for f32/f16/i8
dot products with f32 accumulation; NEST_FORCE_SCALAR=1 disables for
before/after benches. Optional sections 0x07 (HNSW) and 0x08 (BM25)
plus exact rerank for ANN/hybrid so the final score remains real
cosine.

CLI: search-text (embeds via python/embed_query.py and honors the
manifest's index_type), search-ann --ef, benchmark --ann <ef> with
recall@k vs exact. PyO3 + builder.py expose presets exact /
compressed / tiny / hybrid.

Acceptance on data/corpus_next.v1.nest (30,725 chunks, dim 384):
  exact      119.94 MB  recall=1.0000  p50=3.41ms
  compressed  41.99 MB  recall=1.0000  p50=3.88ms  drift_max=4e-5
  tiny        34.64 MB  recall=0.9733  p50=0.69ms  drift_max=8e-4
  hybrid      80.88 MB  recall=0.9933  p50=8.28ms  drift_max=0
SIMD on/off: 3.12ms vs 9.34ms exact (~3x speedup, NEON).

cargo test --workspace: 94/94 pass; clippy + fmt clean; golden
fixture hashes preserved.

+++ crates/nest-format/tests/roundtrip.rs
@@ -424,7 +424,7 @@ fn citation_id_uses_content_hash_with_nest_scheme() {
 #[test]
 fn rejects_unsupported_dtype_in_manifest() {
     let mut m = manifest(4, 1);
-    m.dtype = "float16".into();
+    m.dtype = "bfloat16".into();
     let res = NestFileBuilder::new(m)
         .add_chunk(chunk("a", "b", 0, 1, vec![1.0, 0.0, 0.0, 0.0]))
         .build_bytes();
